Fix object can't be converted to string on subscription renewal
When a subscription can't be renewed because of some card or funds issue,
the response property of the API result is an array, instead of a string.
This causes issues when it's used in a sprintf to add the order note.

This commit fixes that by checking if "response" is an array and using
the appropriate values for each case.